migrate ChangeColumn out of /changes/new
